// src/generated-types/menu-category.ts
export interface MenuCategory {
  id: string;
  slug: string;
  position: number;
  imageUrl?: string | null | undefined;
  isAvailable: boolean;
  createdAt: Date | null;
  updatedAt: Date | null;
}

/** Message representing a menu category translation */
export interface MenuCategoryTranslation {
  id: string;
  language: string;
  title: string;
  description?: string | null | undefined;
}

/** Message representing a menu category with translation */
export interface MenuCategoryWithTranslation {
  id: string;
  slug: string;
  position: number;
  imageUrl?: string | null | undefined;
  isAvailable: boolean;
  translations: MenuCategoryTranslation[];
}

/** Message representing a menu category with items */
export interface MenuCategoryWithItems {
  id: string;
  slug: string;
  position: number;
  imageUrl?: string | null | undefined;
  isAvailable: boolean;
  translations: MenuCategoryTranslation[];
  menuItems: MenuItemWithTranslation[];
}